Let's try to use Neovide for a bit. I know I loose features from Tmux
- selecting everything I see on a screen
- open editor in the background

I want to try Neovide just because of better text and color rendering,
so yeah. It's slightly better. I also like the pleasant animation of
curson moving as well as the existing ability to have padding on the left
and right, so I can have kind of centered text view and kind of zen
mode.

This is still an experiment, not sure I will stick to it, but for now
only thing I can say it's: Let's gooooooo!